Add expenses_manage_access checking in Expenses views and template

+++ expenses/views.py
@@ -22,47 +22,50 @@ class Expenses:
     @staticmethod
     @login_required
     def add_expense(request):
-        if request.method.lower() == 'post':
-            project = Project.objects.get(name=request.POST.get('project'), company=request.user.company)
-            category = Category.objects.get(name=request.POST.get('category'), company=request.user.company)
-            notes = request.POST.get('notes')
-            amount = request.POST.get('amount')
-            
-            expense = Expense.objects.create(
-                company=request.user.company,
-                project=project,
-                category=category,
-                notes=notes,
-                amount=amount
-            )
-            expense.save()
+        if request.user.role.expenses_manage_access:
+            if request.method.lower() == 'post':
+                project = Project.objects.get(name=request.POST.get('project'), company=request.user.company)
+                category = Category.objects.get(name=request.POST.get('category'), company=request.user.company)
+                notes = request.POST.get('notes')
+                amount = request.POST.get('amount')
+                
+                expense = Expense.objects.create(
+                    company=request.user.company,
+                    project=project,
+                    category=category,
+                    notes=notes,
+                    amount=amount
+                )
+                expense.save()
 
-            project.total_spent += float(amount)
-            project.save()
+                project.total_spent += float(amount)
+                project.save()
         return redirect('expenses')
 
     @staticmethod
     @login_required
     def edit_expense(request, expense_id):
-        if request.method.lower() == 'post':
-            expense = Expense.objects.get(pk=expense_id, company=request.user.company)
-            project = Project.objects.get(name=request.POST.get('project'), company=request.user.company)
-            category = Category.objects.get(name=request.POST.get('category'), company=request.user.company)
-            notes = request.POST.get('notes')
-            amount = request.POST.get('amount')
+        if request.user.role.expenses_manage_access:
+            if request.method.lower() == 'post':
+                expense = Expense.objects.get(pk=expense_id, company=request.user.company)
+                project = Project.objects.get(name=request.POST.get('project'), company=request.user.company)
+                category = Category.objects.get(name=request.POST.get('category'), company=request.user.company)
+                notes = request.POST.get('notes')
+                amount = request.POST.get('amount')
 
-            expense.project = project
-            expense.category = category
-            expense.notes = notes
-            expense.amount = amount
-            expense.save()
+                expense.project = project
+                expense.category = category
+                expense.notes = notes
+                expense.amount = amount
+                expense.save()
         return redirect('expenses')
 
     @staticmethod
     @login_required
     def delete_expense(request, expense_id):
-        expense = Expense.objects.get(pk=expense_id, company=request.user.company)
-        expense.project.total_spent -= expense.amount
-        expense.project.save()
-        expense.delete()
+        if request.user.role.expenses_manage_access:
+            expense = Expense.objects.get(pk=expense_id, company=request.user.company)
+            expense.project.total_spent -= expense.amount
+            expense.project.save()
+            expense.delete()
         return redirect('expenses')
